Errors in mapping fixed

+++ src/main/java/pl/lodz/p/it/ssbd2024/ssbd01/entities/util/AbstractEntity.java
@@ -19,5 +19,5 @@ public abstract class AbstractEntity {
     private UUID id;
 
     @Version
-    private BigInteger version;
+    private Integer version;
 }
